support use highlighter with highlight query
