+++ internal/proxy/search_pipeline_test.go
@@ -299,8 +299,8 @@ func (s *SearchPipelineSuite) TestHighlightOp() {
 		DbName:         "default",
 	}
 
-	highlightTasks := []*highlightTask{
-		{
+	highlightTasks := map[int64]*highlightTask{
+		100: {
 			HighlightTask: &querypb.HighlightTask{
 				Texts:     []string{"target text"},
 				FieldName: testVarCharField,
@@ -313,8 +313,10 @@ func (s *SearchPipelineSuite) TestHighlightOp() {
 
 	mockLb := shardclient.NewMockLBPolicy(s.T())
 	searchTask := &searchTask{
-		node:           proxy,
-		highlightTasks: highlightTasks,
+		node: proxy,
+		highlighter: &LexicalHighlighter{
+			tasks: highlightTasks,
+		},
 		lb:             mockLb,
 		schema:         newSchemaInfo(schema),
 		request:        req,
